+++ conda/cli/help.py
@@ -37,7 +37,7 @@ Missing write permissions in: ${root_dir}
 #
 # $ conda create -n my_${name} --clone=${prefix}
 """
-    msg = msg.replace('${root_dir}', context.root_dir)
+    msg = msg.replace('${root_dir}', context.root_prefix)
     msg = msg.replace('${prefix}', prefix)
     msg = msg.replace('${name}', name_prefix(prefix))
     msg = msg.replace('${command}', command)