+++ conda/common/signals.py
@@ -7,6 +7,13 @@ from .compat import iteritems
 
 
 def get_signal_name(signum):
+    """
+    Examples:
+        >>> from signal import SIGINT
+        >>> get_signal_name(SIGINT)
+        'SIGINT'
+
+    """
     return next((k for k, v in iteritems(signal.__dict__)
                  if v == signum and k.startswith('SIG') and not k.startswith('SIG_')),
                 None)