+++ conda/plan.py
@@ -11,10 +11,10 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 from collections import defaultdict, namedtuple
 from logging import getLogger
-from os.path import abspath, basename, exists, isdir, join
+from os.path import abspath, basename, exists, join
 import sys
 
-from . import CondaError, instructions as inst
+from . import instructions as inst
 from ._vendor.boltons.setutils import IndexedSet
 from .base.constants import DEFAULTS_CHANNEL_NAME, UNKNOWN_CHANNEL
 from .base.context import context
@@ -28,7 +28,6 @@ from .core.linked_data import is_linked, linked_data
 from .core.package_cache import ProgressiveFetchExtract
 from .exceptions import (ArgumentError, CondaIndexError, CondaRuntimeError,
                          InstallError, RemoveError)
-from .gateways.disk.create import mkdir_p
 from .history import History
 from .instructions import (ACTION_CODES, CHECK_EXTRACT, CHECK_FETCH, EXTRACT, FETCH, LINK, PREFIX,
                            PRINT, PROGRESS, PROGRESSIVEFETCHEXTRACT, PROGRESS_COMMANDS,
@@ -316,16 +315,6 @@ def inject_UNLINKLINKTRANSACTION(plan, index, prefix):
         link_dists = tuple(Dist(d[1]) for d in grouped_instructions.get(LINK, ()))
         unlink_dists, link_dists = handle_menuinst(unlink_dists, link_dists)
 
-        # make sure prefix directory exists
-        if link_dists:
-            if not isdir(prefix):
-                try:
-                    mkdir_p(prefix)
-                except (IOError, OSError) as e:
-                    log.debug(repr(e))
-                    raise CondaError("Unable to create prefix directory '%s'.\n"
-                                     "Check that you have sufficient permissions." % prefix)
-
         # TODO: ideally we'd move these two lines before both the y/n confirmation and the --dry-run exit  # NOQA
         pfe = ProgressiveFetchExtract(index, link_dists)
         pfe.prepare()
@@ -392,19 +381,18 @@ def ensure_linked_actions(dists, prefix, index=None, force=False,
 
 
 def is_root_prefix(prefix):
-    return abspath(prefix) == abspath(context.root_dir)
+    return abspath(prefix) == abspath(context.root_prefix)
 
 
-def add_defaults_to_specs(r, linked, specs, update=False):
-    # TODO: This should use the pinning mechanism. But don't change the API:
-    # cas uses it.
-    if r.explicit(specs):
+def add_defaults_to_specs(r, linked, specs, update=False, prefix=None):
+    # TODO: This should use the pinning mechanism. But don't change the API because cas uses it
+    if r.explicit(specs) or is_private_env(prefix):
         return
     log.debug('H0 specs=%r' % specs)
     names_linked = {r.package_name(d): d for d in linked if d in r.index}
     mspecs = list(map(MatchSpec, specs))
 
-    for name, def_ver in [('python', context.default_python),
+    for name, def_ver in [('python', context.default_python or None),
                           # Default version required, but only used for Python
                           ('lua', None)]:
         if any(s.name == name and not s.is_simple() for s in mspecs):
@@ -439,7 +427,7 @@ def add_defaults_to_specs(r, linked, specs, update=False):
             specs.append(spec)
             continue
 
-        if name == 'python' and def_ver.startswith('3.'):
+        if name == 'python' and def_ver and def_ver.startswith('3.'):
             # Don't include Python 3 in the specs if this is the Python 3
             # version of conda.
             continue
@@ -540,7 +528,7 @@ def add_unlink_options_for_update(actions, required_solves, index):
                     else:
                         actions.append(remove_actions(context.root_prefix, matched_in_root, index))
         # If the solved prefix is root
-        elif preferred_env_matches_prefix(None, solved.prefix, context.root_dir):
+        elif preferred_env_matches_prefix(None, solved.prefix, context.root_prefix):
             for spec in solved.specs:
                 spec_in_private_env = prefix_if_in_private_env(spec)
                 if spec_in_private_env:
@@ -583,8 +571,8 @@ def ensure_packge_not_duplicated_in_private_env_root(dists_for_envs, linked_in_r
 
 
 def not_requires_private_env(prefix, preferred_envs):
-    if (context.prefix_specified is True or not context.prefix == context.root_dir or
-            all(preferred_env_matches_prefix(preferred_env, prefix, context.root_dir) for
+    if (context.prefix_specified is True or not context.prefix == context.root_prefix or
+            all(preferred_env_matches_prefix(preferred_env, prefix, context.root_prefix) for
                 preferred_env in preferred_envs)):
         return True
     return False
@@ -605,22 +593,22 @@ def determine_dists_per_prefix(r, prefix, index, preferred_envs, dists_for_envs,
         prefix_with_dists_no_deps_has_resolve = [SpecsForPrefix(prefix=prefix, r=r, specs=dists)]
     else:
         # Ensure that conda is working in the root dir
-        assert(context.prefix == context.root_dir)
+        assert context.prefix == context.root_prefix
 
         def get_r(preferred_env):
             # don't make r for the prefix where we already have it created
-            if preferred_env_matches_prefix(preferred_env, prefix, context.root_dir):
+            if preferred_env_matches_prefix(preferred_env, prefix, context.root_prefix):
                 return r
             else:
                 return get_resolve_object(index.copy(), preferred_env_to_prefix(
-                    preferred_env, context.root_dir, context.envs_dirs))
+                    preferred_env, context.root_prefix, context.envs_dirs))
 
         prefix_with_dists_no_deps_has_resolve = []
         for env in preferred_envs:
             dists = IndexedSet(d.spec for d in dists_for_envs if d.env == env)
             prefix_with_dists_no_deps_has_resolve.append(
                 SpecsForPrefix(
-                    prefix=preferred_env_to_prefix(env, context.root_dir, context.envs_dirs),
+                    prefix=preferred_env_to_prefix(env, context.root_prefix, context.envs_dirs),
                     r=get_r(env),
                     specs=dists)
             )
@@ -638,7 +626,7 @@ def match_to_original_specs(str_specs, specs_for_prefix):
             matched = matches_any_spec(spec)
             if matched:
                 new_matches.append(matched)
-        add_defaults_to_specs(r, linked, new_matches)
+        add_defaults_to_specs(r, linked, new_matches, prefix=prefix_with_dists.prefix)
         matched_specs_for_prefix.append(SpecsForPrefix(
             prefix=prefix_with_dists.prefix, r=prefix_with_dists.r, specs=new_matches))
     return matched_specs_for_prefix
@@ -703,7 +691,7 @@ These packages need to be removed before conda can proceed.""" % (' '.join(linke
         force=force, always_copy=always_copy)
 
     if actions[LINK]:
-        actions[SYMLINK_CONDA] = [context.root_dir]
+        actions[SYMLINK_CONDA] = [context.root_prefix]
 
     for dist in sorted(linked):
         dist = Dist(dist)